Add pack/unpack commands
Pack and Unpack commands based losely on php/Perl pack/unpack functions.
Only format codes n, N, a, x are supported. 'a' can take a length of * to indicate a variable length string. You will need to add x (for a null byte) after it if this is the case and this is looked for specifically. e.g. a*x counts as one decode entry. currently x is not supported as an individual decode format.